add dashboard

+++ app/Http/Controllers/JobController.php
@@ -2,19 +2,26 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Http\Request;
 use Illuminate\View\View;
-use App\Models\Job;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Storage;
-
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use App\Models\Job;
 class JobController extends Controller
 {
+
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
+    // @desc Show all job listings
+    // @route GET /jobs
     public function index(): View
     {
+
         $jobs = Job::all();
 
         return view('jobs.index')->with('jobs', $jobs);
@@ -23,14 +30,20 @@ class JobController extends Controller
     /**
      * Show the form for creating a new resource.
      */
-    public function create(): View
+
+
+    public function create()
     {
+
+
         return view('jobs.create');
     }
 
     /**
      * Store a newly created resource in storage.
      */
+    // @desc Save job to database
+    // @route POST /jobs
     public function store(Request $request): RedirectResponse
     {
 
@@ -55,7 +68,7 @@ class JobController extends Controller
             'company_website' => 'nullable|url',
         ]);
 
-        $validatedData['user_id'] = 1;
+        $validatedData['user_id'] = auth()->user()->id;
 
         // Check for image
         if($request->hasFile('company_logo')) {
@@ -77,6 +90,8 @@ class JobController extends Controller
     /**
      * Display the specified resource.
      */
+    // @desc Display a single job listing
+    // @route GET /jobs/{$id}
     public function show(job $job): View
     {
         return view('jobs.show')->with('job', $job);
@@ -85,16 +100,24 @@ class JobController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
+    // @desc Show edit job form
+    // @route GET /login
     public function edit(Job $job): View
     {
+        $this->authorize('update', $job);
         return view('jobs.edit')->with('job', $job);
     }
 
     /**
      * Update the specified resource in storage.
      */
+    // @desc Update job listing
+    // @route PUT /jobs/{$id}
     public function update(Request $request, Job $job): string
     {
+        // Check if user is authorized
+        $this->authorize('update', $job);
+
         $validatedData = $request->validate([
             'title' => 'required|max:255',
             'description' => 'required|string',
@@ -140,14 +163,25 @@ class JobController extends Controller
     /**
      * Remove the specified resource from storage.
      */
+    // @desc Delete a job listing
+    // @route DELETE /jobs/{$id}
     public function destroy(Job $job): RedirectResponse
     {
+
+        $this->authorize('delete', $job);
+
         // If logo, then delete it
         if($job->company_logo) {
             Storage::delete('public/logos/' .$job->company_logo);
         }
 
         $job->delete();
+
+        // Check if request came from the dashboard
+        if(request()->query('from') == 'dashboard') {
+            return redirect()->route('dashboard')->with('success', 'Job listing deleted successfully.');
+        }
+
         return redirect()->route('jobs.index')->with('success', 'Job listing deleted successfully.');
     }
